requirements.txt
pyyaml
flask
sqlalchemy
flask-cors